requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cors==1.0.1
distlib==0.3.4
filelock==3.7.0
Flask==2.1.2
Flask-Cors==3.0.10
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.1.1
platformdirs==2.5.2
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.27.1
requests-file==1.5.1
six==1.16.0
tldextract==3.3.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.2
zope.event==4.5.0
zope.interface==5.4.0
